third_party/blink/web_tests/images/animated-images-use-counter-002.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<body>
<script>
const kAnimatedImageUsedMoreThanOnce = 5605;

async_test(function(t) {
  internals.clearUseCounter(document, kAnimatedImageUsedMoreThanOnce);

  let gif1 = document.createElement("img");
  gif1.src = "resources/animated.gif";
  document.body.appendChild(gif1);

  onload = t.step_func(() => {
    requestAnimationFrame(t.step_func(() => {
      requestAnimationFrame(t.step_func_done(() => {
        assert_false(internals.isUseCounted(document, kAnimatedImageUsedMoreThanOnce));
      }));
    }));
  });
}, 'Check that kAnimatedImageUsedMoreThanOnce is NOT counted with a single animated image');
</script>
</body>
